refactor: move all build shim logic into the shared layer
