+++ src/obsidian/dashboards-embedded-render-child.tsx
@@ -2,11 +2,11 @@ import { MarkdownRenderChild, TFile } from "obsidian";
 
 import "react-devtools";
 import { Root, createRoot } from "react-dom/client";
-import Dashboard from "src/obsidian-shim/build/dashboard";
 import { store } from "src/redux/global/store";
 import DashboardsView from "./dashboards-view";
 import { deserializeTableState } from "src/data/serialize-table-state";
 import { v4 as uuidv4 } from "uuid";
+import DashboardsApp from "src/react/table-app";
 
 export default class DashboardsEmbeddedRenderChild extends MarkdownRenderChild {
 	private root: Root | null;
@@ -38,7 +38,7 @@ export default class DashboardsEmbeddedRenderChild extends MarkdownRenderChild {
 			this.root = createRoot(container);
 
 			this.root.render(
-				<Dashboard
+				<DashboardsApp
 					leaf={activeView.leaf}
 					appId={this.appId}
 					filePath={file.path}
