+++ src/react/shared/icon/index.tsx
@@ -1,6 +1,6 @@
+import { setIcon } from "obsidian";
 import React from "react";
 
-import { setIcon } from "src/obsidian-shim/build/set-icon";
 import { appendOrReplaceFirstChild } from "src/shared/render/utils";
 
 interface Props {
